+++ NEWS.md
@@ -1,3 +1,10 @@
+# muiMaterial 0.1.2
+
+- improve examples
+- `muiMaterialPage()` gives access to Google icons (symbols) and Roboto font.
+- export `muiMaterialDependency()`
+- improve internal file structure for ThemeProvider and muiMaterialDependency.
+
 # muiMaterial 0.1.0
 
-* initial commit
+- initial commit